macro_detection_program/docker_images/mk_macro_program/requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
tensorflow==2.17.0
memory-profiler==0.61.0
